Disable status updates by decoding Hatch state characteristic
The Hatch seems to stay connected after you close the phone app after
using it, and it prevents the Hatch controller from connecting to it for
status updates, which causes the heartbeat alarm to fire when it gets
stuck in a retry loop. Those updates aren't really useful anyway, now
that we're mostly just relying on the Hatch scheduled programs these
days.